All compression tests compile with vs2015, compression lib

+++ lib/compression/test/test_compression_lzham.cpp
@@ -38,7 +38,7 @@ int main(int argc, char *argv[])
   foreach(data, i)
     if (data[i] != uncompressed[i])
     {
-      vsx_printf(L"data differs on index %d\n", i);
+      vsx_printf(L"data differs on index %d\n", (int)i);
       exit(1);
     }
